Decompose toolpath overlay
